Move `ArticleFormat` decision functions into `format.ts`
Consolidates the types and core functions that relate to `ArticleFormat` in one file.
