+++ dotcom-rendering/src/components/RichLinkComponent.importable.tsx
@@ -1,5 +1,4 @@
-import { decideFormat } from '../lib/decideFormat';
-import { type ArticleFormat } from '../lib/format';
+import { type ArticleFormat, decideFormat } from '../lib/format';
 import { useApi } from '../lib/useApi';
 import type { RichLinkBlockElement, StarRating } from '../types/content';
 import type { FEFormat } from '../types/frontend';
