+++ dotcom-rendering/src/layouts/DecideLayout.stories.tsx
@@ -15,8 +15,7 @@ import { Recipe as RecipeStandardLifestyleFixture } from '../../fixtures/generat
 import { Standard as StandardStandardNewsFixture } from '../../fixtures/generated/dcr-articles/Standard';
 import { Video as VideoStandardNewsFixture } from '../../fixtures/generated/dcr-articles/Video';
 import { embedIframe } from '../client/embedIframe';
-import { decideFormat } from '../lib/decideFormat';
-import type { ArticleFormat } from '../lib/format';
+import { type ArticleFormat, decideFormat } from '../lib/format';
 import { getCurrentPillar } from '../lib/layoutHelpers';
 import { mockRESTCalls } from '../lib/mockRESTCalls';
 import { extractNAV } from '../model/extract-nav';
