--- dotcom-rendering/src/lib/decideFormat.ts
@@ -1,11 +1,0 @@
-import type { ArticleFormat } from '../lib/format';
-import type { FEFormat } from '../types/frontend';
-import { decideDesign } from './decideDesign';
-import { decideDisplay } from './decideDisplay';
-import { decideTheme } from './decideTheme';
-
-export const decideFormat = (format: Partial<FEFormat>): ArticleFormat => ({
-	display: decideDisplay(format),
-	theme: decideTheme(format),
-	design: decideDesign(format),
-});
