--- dotcom-rendering/src/lib/decideTheme.ts
@@ -1,25 +1,0 @@
-import { ArticleSpecial, type ArticleTheme, Pillar } from '../lib/format';
-import type { FEFormat } from '../types/frontend';
-
-export const decideTheme = ({ theme }: Partial<FEFormat>): ArticleTheme => {
-	switch (theme) {
-		case 'NewsPillar':
-			return Pillar.News;
-		case 'OpinionPillar':
-			return Pillar.Opinion;
-		case 'SportPillar':
-			return Pillar.Sport;
-		case 'CulturePillar':
-			return Pillar.Culture;
-		case 'LifestylePillar':
-			return Pillar.Lifestyle;
-		case 'SpecialReportTheme':
-			return ArticleSpecial.SpecialReport;
-		case 'SpecialReportAltTheme':
-			return ArticleSpecial.SpecialReportAlt;
-		case 'Labs':
-			return ArticleSpecial.Labs;
-		default:
-			return Pillar.News;
-	}
-};
