+++ dotcom-rendering/src/lib/format.ts
@@ -59,6 +59,117 @@ export interface ArticleFormat {
 	display: ArticleDisplay;
 }
 
+/**
+ * NOTE: Immersive Opinion pieces are not supported,
+ * i.e. when `CommentDesign` and `ImmersiveDisplay` are used jointly.
+ */
+export const decideDesign = ({ design }: Partial<FEFormat>): ArticleDesign => {
+	switch (design) {
+		case 'ArticleDesign':
+			return ArticleDesign.Standard;
+		case 'PictureDesign':
+			return ArticleDesign.Picture;
+		case 'GalleryDesign':
+			return ArticleDesign.Gallery;
+		case 'AudioDesign':
+			return ArticleDesign.Audio;
+		case 'VideoDesign':
+			return ArticleDesign.Video;
+		case 'ReviewDesign':
+			return ArticleDesign.Review;
+		case 'AnalysisDesign':
+			return ArticleDesign.Analysis;
+		case 'CommentDesign':
+			return ArticleDesign.Comment;
+		case 'LetterDesign':
+			return ArticleDesign.Letter;
+		case 'FeatureDesign':
+			return ArticleDesign.Feature;
+		case 'LiveBlogDesign':
+			return ArticleDesign.LiveBlog;
+		case 'DeadBlogDesign':
+			return ArticleDesign.DeadBlog;
+		case 'RecipeDesign':
+			return ArticleDesign.Recipe;
+		case 'MatchReportDesign':
+			return ArticleDesign.MatchReport;
+		case 'InterviewDesign':
+			return ArticleDesign.Interview;
+		case 'EditorialDesign':
+			return ArticleDesign.Editorial;
+		case 'QuizDesign':
+			return ArticleDesign.Quiz;
+		case 'InteractiveDesign':
+			return ArticleDesign.Interactive;
+		case 'PhotoEssayDesign':
+			return ArticleDesign.PhotoEssay;
+		case 'ObituaryDesign':
+			return ArticleDesign.Obituary;
+		case 'FullPageInteractiveDesign':
+			return ArticleDesign.FullPageInteractive;
+		case 'NewsletterSignupDesign':
+			return ArticleDesign.NewsletterSignup;
+		case 'ExplainerDesign':
+			return ArticleDesign.Explainer;
+		case 'TimelineDesign':
+			return ArticleDesign.Timeline;
+		case 'ProfileDesign':
+			return ArticleDesign.Profile;
+		default:
+			return ArticleDesign.Standard;
+	}
+};
+
+const decideDisplay = ({
+	display,
+	design,
+}: Partial<FEFormat>): ArticleDisplay => {
+	switch (display) {
+		case 'StandardDisplay':
+			return ArticleDisplay.Standard;
+		case 'ImmersiveDisplay':
+			// Temporary hack until we can handle Immersive Comment pieces
+			return design === 'CommentDesign'
+				? ArticleDisplay.Standard
+				: ArticleDisplay.Immersive;
+		case 'ShowcaseDisplay':
+			return ArticleDisplay.Showcase;
+		case 'NumberedListDisplay':
+			return ArticleDisplay.NumberedList;
+		default:
+			return ArticleDisplay.Standard;
+	}
+};
+
+export const decideTheme = ({ theme }: Partial<FEFormat>): ArticleTheme => {
+	switch (theme) {
+		case 'NewsPillar':
+			return Pillar.News;
+		case 'OpinionPillar':
+			return Pillar.Opinion;
+		case 'SportPillar':
+			return Pillar.Sport;
+		case 'CulturePillar':
+			return Pillar.Culture;
+		case 'LifestylePillar':
+			return Pillar.Lifestyle;
+		case 'SpecialReportTheme':
+			return ArticleSpecial.SpecialReport;
+		case 'SpecialReportAltTheme':
+			return ArticleSpecial.SpecialReportAlt;
+		case 'Labs':
+			return ArticleSpecial.Labs;
+		default:
+			return Pillar.News;
+	}
+};
+
+export const decideFormat = (format: Partial<FEFormat>): ArticleFormat => ({
+	design: decideDesign(format),
+	display: decideDisplay(format),
+	theme: decideTheme(format),
+});
+
 export const getThemeNameAsString = (format: ArticleFormat): string => {
 	const themeName = Pillar[format.theme] ?? ArticleSpecial[format.theme];
 	if (!themeName) throw new Error('Unknown theme');
