+++ dotcom-rendering/src/server/handler.article.amp.tsx
@@ -4,10 +4,9 @@ import type { AnalyticsModel } from '../components/Analytics.amp';
 import { AmpArticlePage } from '../components/ArticlePage.amp';
 import { isAmpSupported } from '../components/Elements.amp';
 import type { PermutiveModel } from '../components/Permutive.amp';
-import { decideFormat } from '../lib/decideFormat';
 import { enhance } from '../lib/enhance.amp';
 import { NotRenderableInDCR } from '../lib/errors/not-renderable-in-dcr';
-import { ArticleDesign } from '../lib/format';
+import { ArticleDesign, decideFormat } from '../lib/format';
 import { generatePermutivePayload } from '../lib/permutive.amp';
 import { extractScripts } from '../lib/scripts.amp';
 import { findBySubsection } from '../model/article-sections';
